dynamic object changes
- added :box type to physics objects for dynamic object's
  load-physics-objects default method.
- consolidated level objects into (level-objects level). there is no
  more concept of level actors. all level objects are stored in that one
  list. this also means some queries had to be changed, specifically in
  init-level-physics-objects (now manually selecting level-object
  objects instead of come-one-come-all).
- ignoring ALL collisions in event debug output. was too noisy/slow.
- no longer updating actor state in world's step-game-world fn. this is
  for the game itself to do, not the engine (and apparently the demo
  isnt even using this code newayz)

there are probably a few more changes, but that's the brunt of it

add a new dynamic object (platform) to the demo's tree level. looks
stupid right now, but deomstrates the ability to add dynamic objects!
yay! next step is to get the platform to move when user is on it, mario
style.