renaming game-object
game-object is being renamed to base-object. i think this makes more
sense as ghostie gets more high level, because many objects (actors,
dynamics objects, etc) are built on top of it.

not only this, but ALL game-object (now base-object) accessor have been
renamed to rename the game- (base-) prefix:

`game-object-physics-body` is now `object-physics-body`.

this is a lot less typing, and a lot more clear.